refactor: Update MainScreen layout and add header with new todo button
This commit refactors the MainScreen layout by using a Padding widget to add padding around the main content. It also introduces a new header section with a button for adding new todos. The header is implemented as a separate widget (_Header) to improve code organization and reusability.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring.
